// types/global.ts
export interface Translation {
  id: number;
  locale: string;
  title: string;
}
// ...
export interface Country {
  img: string;
  id: number;
  active: boolean;
  region_id: number;
  translation: Translation | null;
  cities_count: number;
  code: string;
  // eslint-disable-next-line no-use-before-define
  city?: City | null | undefined;
}

export interface City extends Country {
  country_id: number;
}